+++ src/FSharp.AWS.DynamoDB/RecordKeySchema.fs
@@ -368,12 +368,9 @@ type TableKeySchemata with
                 yield! td.GlobalSecondaryIndexes |> Seq.map mkGlobalSecondaryIndex
                 yield! td.LocalSecondaryIndexes |> Seq.map mkLocalSecondaryIndex |])
 
-    /// Create a CreateTableRequest using supplied key schema
-    member schema.CreateCreateTableRequest (tableName : string, provisionedThroughput : ProvisionedThroughput) =
-        let ctr = CreateTableRequest(TableName = tableName)
+    /// Applies the settings implied by the schema to the supplied CreateTableRequest
+    member schema.ApplyToCreateTableRequest(ctr : CreateTableRequest) =
         let inline mkKSE n t = KeySchemaElement(n, t)
-
-        ctr.ProvisionedThroughput <- provisionedThroughput
 
         let keyAttrs = new Dictionary<string, KeyAttributeSchema>()
         for tks in schema.Schemata do
@@ -391,7 +388,6 @@ type TableKeySchemata with
                 gsi.KeySchema.Add <| mkKSE tks.HashKey.AttributeName KeyType.HASH
                 tks.RangeKey |> Option.iter (fun rk -> gsi.KeySchema.Add <| mkKSE rk.AttributeName KeyType.RANGE)
                 gsi.Projection <- Projection(ProjectionType = ProjectionType.ALL)
-                gsi.ProvisionedThroughput <- provisionedThroughput
                 ctr.GlobalSecondaryIndexes.Add gsi
 
             | LocalSecondaryIndex name ->
@@ -405,5 +401,3 @@ type TableKeySchemata with
         for attr in keyAttrs.Values do
             let ad = AttributeDefinition(attr.AttributeName, attr.KeyType)
             ctr.AttributeDefinitions.Add ad
-
-        ctr