+++ tests/FSharp.AWS.DynamoDB.Tests/MetricsCollectorTests.fs
@@ -52,7 +52,7 @@ let (|TotalCu|) : ConsumedCapacity list -> float = Seq.sumBy (fun c -> c.Capacit
 /// Tests without common setup
 type Tests(fixture : TableFixture) =
 
-    let rawTable = TableContext.Create<MetricsRecord>(fixture.Client, fixture.TableName, createIfNotExists = true)
+    let rawTable = fixture.CreateEmpty<MetricsRecord>()
 
     let (|ExpectedTableName|_|) name = if name = fixture.TableName then Some () else None
 
@@ -84,7 +84,7 @@ type Tests(fixture : TableFixture) =
 /// Tests that look up a specific item. Each test run gets a fresh individual item
 type ItemTests(fixture : TableFixture) =
 
-    let rawTable = TableContext.Create<MetricsRecord>(fixture.Client, fixture.TableName, createIfNotExists = true)
+    let rawTable = fixture.CreateEmpty<MetricsRecord>()
     let (|ExpectedTableName|_|) name = if name = fixture.TableName then Some () else None
 
     let item = mkItem (guid()) (guid()) 0
@@ -128,10 +128,10 @@ type ItemTests(fixture : TableFixture) =
 
     interface IClassFixture<TableFixture>
 
-/// Heavy tests reliant on establishing (and mutating) multiple items. Separate Test Class so Xunit will run thme in parallel with others
+/// Heavy tests reliant on establishing (and mutating) multiple items. Separate Test Class so Xunit will run them in parallel with others
 type BulkMutationTests(fixture : TableFixture) =
 
-    let rawTable = TableContext.Create<MetricsRecord>(fixture.Client, fixture.TableName, createIfNotExists = true)
+    let rawTable = fixture.CreateEmpty<MetricsRecord>()
     let (|ExpectedTableName|_|) name = if name = fixture.TableName then Some () else None
 
     // NOTE we mutate the items so they need to be established each time
@@ -169,7 +169,7 @@ type ManyReadOnlyItemsFixture() =
     inherit TableFixture()
 
     // TOCONSIDER shift this into IAsyncLifetime.InitializeAsync
-    let table = TableContext.Create<MetricsRecord>(base.Client, base.TableName, createIfNotExists = true)
+    let table = base.CreateEmpty<MetricsRecord>()
 
     let hk = guid ()
     do  let gsk = guid ()