+++ tests/FSharp.AWS.DynamoDB.Tests/SparseGSITests.fs
@@ -31,7 +31,7 @@ type ``Sparse GSI Tests`` (fixture : TableFixture) =
             SecondaryHashKey = if rand() % 2L = 0L then Some (guid()) else None ;
         }
 
-    let table = TableContext.Create<GsiRecord>(fixture.Client, fixture.TableName, createIfNotExists = true)
+    let table = fixture.CreateEmpty<GsiRecord>()
 
     let [<Fact>] ``GSI Put Operation`` () =
         let value = mkItem()