+++ tests/FSharp.AWS.DynamoDB.Tests/Utils.fs
@@ -5,18 +5,20 @@ open System.IO
 
 open FsCheck
 open Swensen.Unquote
+open Xunit
+
+open FSharp.AWS.DynamoDB
 
 open Amazon.DynamoDBv2
 open Amazon.Runtime
-open Xunit
 
 [<AutoOpen>]
 module Utils =
 
-    let getRandomTableName() =
-        sprintf "fsdynamodb-%s" <| Guid.NewGuid().ToString("N")
+    let guid () = Guid.NewGuid().ToString("N")
 
-    let guid() = Guid.NewGuid().ToString("N")
+    let getRandomTableName() =
+        sprintf "fsdynamodb-%s" <| guid ()
 
     let shouldFailwith<'T, 'Exn when 'Exn :> exn>(f : unit -> 'T) =
         <@ f () |>  ignore @>
@@ -24,8 +26,8 @@ module Utils =
 
     let getDynamoDBAccount () =
         let credentials = BasicAWSCredentials("Fake", "Fake")
-        let config = AmazonDynamoDBConfig()
-        config.ServiceURL <- "http://localhost:8000"
+        let config = AmazonDynamoDBConfig(ServiceURL = "http://localhost:8000")
+
         new AmazonDynamoDBClient(credentials, config) :> IAmazonDynamoDB
 
 
@@ -37,14 +39,19 @@ module Utils =
 
 
     type TableFixture() =
+
         let client = getDynamoDBAccount()
         let tableName = getRandomTableName()
+
         member _.Client = client
         member _.TableName = tableName
+
+        member _.CreateEmpty<'TRecord>() =
+            let throughput = ProvisionedThroughput(readCapacityUnits = 10L, writeCapacityUnits = 10L)
+            Scripting.TableContext.Initialize<'TRecord>(client, tableName, Throughput.Provisioned throughput)
 
         interface IAsyncLifetime with
             member _.InitializeAsync() =
                 System.Threading.Tasks.Task.CompletedTask
             member _.DisposeAsync() =
                 client.DeleteTableAsync(tableName)
-